+++ src/components/UI/Common/Button/Button.tsx
@@ -1,47 +1,48 @@
-import { ReactNode } from 'react'
+import {ReactNode} from 'react'
 import styled from 'styled-components'
 
-const StyledButton = styled.button<{isHidden: boolean}>`
-    font-size: 18px;
-    border-radius: 5px;
-    border: 1px #aaa solid;
-    background-color: white;
-    padding: 10px 0;
-    width: 100%;
-    cursor: pointer;
-    transition: transform 100ms ease-in-out;
-    visibility: ${props => props.isHidden ? 'hidden' : 'visible'};
-    &:hover {
-      transform: scale(1.01);
-    }
-    &:disabled {
-      background-color: rgba(255, 255, 255, 0.5);
-    }
-    &:last-of-type {
-      margin-top: 10px;
-    }
+const StyledButton = styled.button<{ isHidden: boolean }>`
+  font-size: 18px;
+  border-radius: 5px;
+  border: none;
+  padding: 15px 0;
+  max-width: 25rem;
+  width: 100%;
+  cursor: pointer;
+  transition: transform 100ms ease-in-out;
+  visibility: ${props => props.isHidden ? 'hidden' : 'visible'};
+  color: whitesmoke;
+  background-color: ${props => props.theme.ui.primaryColor};
+
+  &:disabled {
+    background-color: rgba(255, 255, 255, 0.5);
+  }
+
+  &:active {
+    outline: none;
+  }
 `
 
 export type ButtonProps = {
-    text: string | ReactNode,
-    disabled: boolean,
+    children: string | ReactNode,
+    disabled?: boolean,
     onClick: () => void,
     isHidden?: boolean
 }
 
 const Button = ({
-    text,
-    isHidden,
-    disabled,
-    onClick
-}: ButtonProps) => {
+                    children,
+                    isHidden,
+                    disabled,
+                    onClick
+                }: ButtonProps) => {
     return (
         <StyledButton
             disabled={disabled}
             isHidden={isHidden || false}
             onClick={onClick}
         >
-            {text}
+            {children}
         </StyledButton>
     )
 }